[refactor] RequestedChartLayout: create new UI model
