+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/destinations/agile/AgileUIState.kt
@@ -10,11 +10,12 @@ package com.rwmobi.kunigami.ui.destinations.agile
 import androidx.compose.runtime.Immutable
 import com.rwmobi.kunigami.domain.model.Rate
 import com.rwmobi.kunigami.ui.model.ErrorMessage
+import com.rwmobi.kunigami.ui.model.RequestedChartLayout
 
 @Immutable
 data class AgileUIState(
     val isLoading: Boolean = true,
-    val requestedLayout: AgileScreenLayout = AgileScreenLayout.Portrait,
+    val requestedChartLayout: RequestedChartLayout = RequestedChartLayout.Portrait,
     val rates: List<Rate> = emptyList(),
     val rateRange: ClosedFloatingPointRange<Double> = 0.0..0.0,
     val errorMessages: List<ErrorMessage> = emptyList(),
