+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/destinations/usage/UsageScreen.kt
@@ -30,6 +30,7 @@ import com.rwmobi.kunigami.domain.extensions.toLocalHourMinuteString
 import com.rwmobi.kunigami.ui.components.LoadingScreen
 import com.rwmobi.kunigami.ui.components.ScrollbarMultiplatform
 import com.rwmobi.kunigami.ui.components.koalaplot.VerticalBarChart
+import com.rwmobi.kunigami.ui.model.RequestedChartLayout
 import com.rwmobi.kunigami.ui.theme.getDimension
 import com.rwmobi.kunigami.ui.utils.generateGYRHueColorPalette
 import io.github.koalaplot.core.bar.DefaultVerticalBarPlotEntry
@@ -74,7 +75,7 @@ fun UsageScreen(
                 }
             }
 
-            val labelIndex: Map<Int, Int> = remember(uiState.consumptions, uiState.requestedLayout) {
+            val labelIndex: Map<Int, Int> = remember(uiState.consumptions, uiState.requestedChartLayout) {
                 buildMap {
                     // Generate all possible labels
                     var lastRateValue: Int? = null
@@ -99,16 +100,17 @@ fun UsageScreen(
                 ) {
                     item {
                         BoxWithConstraints {
-                            val constraintModifier = when (uiState.requestedLayout) {
-                                is UsageScreenLayout.Portrait -> {
+                            val constraintModifier = when (uiState.requestedChartLayout) {
+                                is RequestedChartLayout.Portrait -> {
                                     Modifier
                                         .fillMaxWidth()
                                         .aspectRatio(4 / 3f)
                                 }
 
-                                is UsageScreenLayout.LandScape -> {
-                                    Modifier.fillMaxSize()
-                                        .height(uiState.requestedLayout.requestedMaxHeight)
+                                is RequestedChartLayout.LandScape -> {
+                                    Modifier
+                                        .fillMaxWidth()
+                                        .height(uiState.requestedChartLayout.requestedMaxHeight)
                                 }
                             }
