+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/destinations/usage/UsageUIState.kt
@@ -10,11 +10,12 @@ package com.rwmobi.kunigami.ui.destinations.usage
 import androidx.compose.runtime.Immutable
 import com.rwmobi.kunigami.domain.model.Consumption
 import com.rwmobi.kunigami.ui.model.ErrorMessage
+import com.rwmobi.kunigami.ui.model.RequestedChartLayout
 
 @Immutable
 data class UsageUIState(
     val isLoading: Boolean = true,
-    val requestedLayout: UsageScreenLayout = UsageScreenLayout.Portrait,
+    val requestedChartLayout: RequestedChartLayout = RequestedChartLayout.Portrait,
     val consumptions: List<Consumption> = emptyList(),
     val consumptionRange: ClosedFloatingPointRange<Double> = 0.0..0.0,
     val errorMessages: List<ErrorMessage> = emptyList(),
