+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/model/RequestedChartLayout.kt
@@ -5,13 +5,13 @@
  *
  */
 
-package com.rwmobi.kunigami.ui.destinations.agile
+package com.rwmobi.kunigami.ui.model
 
 import androidx.compose.runtime.Immutable
 import androidx.compose.ui.unit.Dp
 
 @Immutable
-sealed interface AgileScreenLayout {
-    data object Portrait : AgileScreenLayout
-    data class LandScape(val requestedMaxHeight: Dp) : AgileScreenLayout
+sealed interface RequestedChartLayout {
+    data object Portrait : RequestedChartLayout
+    data class LandScape(val requestedMaxHeight: Dp) : RequestedChartLayout
 }
